poc/c_knowledge/requirements_c_knowledge.txt
# FastAPI
fastapi>=0.109.0
uvicorn>=0.27.0

# LLM
openai>=1.12.0
tiktoken>=0.5.0
pydantic-ai>=0.0.12

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.24.0

# Redis
redis>=5.0.0

# Scheduler
apscheduler>=3.10.0

# Data Processing
pydantic>=2.5.0
python-dotenv>=1.0.0

# Batch Processor (크롤링)
selenium>=4.15.0
webdriver-manager>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
